requirements.txt
python-dotenv
google-genai
aiofiles